Moved files from media/helpers to site/Helpers. Moved the generation of the administrative sidebar menu from the list view to the component helper. Replaced redundant template generation in the Organizer view with the use of the administrative sidebar menu. Moved the images and js folders to the site folder. Removed the media folder. Alphabetically ordered the functions of the subject model. Added a getTable function to the subject model. Added form relevance protection to subject functions which modify associations. Made the front and back end subject models identical for future work towards namespacing. Merged the front and backend controllers. Removed the context parameter used in the component helper setUp function. Made the date helper globally available. Standardized the access checks by save functions in models associated with the front end with their backend counterparts.

+++ com_thm_organizer/script.php
@@ -86,7 +86,7 @@ class Com_THM_OrganizerInstallerScript
         </script>
         <fieldset id="com_thm_organizer_fieldset" style="border-radius:10px;">
             <legend>
-                <img style="float:none;" src="../media/com_thm_organizer/images/thm_organizer.png"
+                <img style="float:none;" src="../components/com_thm_organizer/images/thm_organizer.png"
                      alt="THM Organizer Logo"/>
             </legend>
             <div style="padding-left:17px;">
@@ -176,15 +176,17 @@ class Com_THM_OrganizerInstallerScript
             }
 
             $mediaFiles = \JFolder::files(JPATH_SITE . '/media/com_thm_organizer');
-
-            foreach ($mediaFiles as $mediaFile) {
-                \JFile::delete(JPATH_SITE . '/media/com_thm_organizer/' . $mediaFile);
+            if (!empty($mediaFiles)) {
+                foreach ($mediaFiles as $mediaFile) {
+                    \JFile::delete(JPATH_SITE . '/media/com_thm_organizer/' . $mediaFile);
+                }
             }
 
             $mediaFolders = \JFolder::folders(JPATH_SITE . '/media/com_thm_organizer');
-
-            foreach ($mediaFolders as $mediaFolder) {
-                \JFolder::delete(JPATH_SITE . '/media/com_thm_organizer/' . $mediaFolder);
+            if (!empty($mediaFolders)) {
+                foreach ($mediaFolders as $mediaFolder) {
+                    \JFolder::delete(JPATH_SITE . '/media/com_thm_organizer/' . $mediaFolder);
+                }
             }
         }
     }
@@ -215,7 +217,7 @@ class Com_THM_OrganizerInstallerScript
      */
     public function update($parent)
     {
-        $logoURL     = 'media/com_thm_organizer/images/thm_organizer.png';
+        $logoURL     = 'components/com_thm_organizer/images/thm_organizer.png';
         $licenseLink = '<a href="http://www.gnu.org/licenses/gpl-2.0.html" target="_blank">GNU General Public License</a>';
         $version     = (string)$parent->get('manifest')->version;
